requirements.txt
librosa==0.10.1
numpy==1.23.5
concrete-ml==1.3.0
concrete-ml==1.3.0
concrete-python==2.5.0rc1
scikit-learn==1.1.3
scipy==1.10.1
pandas==1.5.3
